Checkpoint save for debugging.
