+++ OpinionsEjb/src/main/java/opca/scraper/TestCAParseSlipDetails.java
@@ -5,15 +5,13 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.logging.Logger;
 
-import opca.model.Disposition;
-import opca.model.PartiesAndAttornies;
 import opca.model.SlipOpinion;
-import opca.model.Summary;
-import opca.model.TrialCourt;
 import opca.parser.ScrapedOpinionDocument;
 
 public class TestCAParseSlipDetails extends CACaseScraper {
+	private static final Logger logger = Logger.getLogger(TestCAParseSlipDetails.class.getName());
 
 	public TestCAParseSlipDetails(boolean debugFiles) {
 		super(debugFiles);
@@ -39,42 +37,54 @@ public class TestCAParseSlipDetails extends CACaseScraper {
 				documents.add(  scrapedDocument );
 				inputStream.close();
 			} catch (IOException e) {
-				e.printStackTrace();
+				logger.severe("Parse Scraped Document: " + e.getMessage());
 			}
 
+			boolean goodtogo = true;
 			try ( InputStream inputStream = Files.newInputStream( 
 					Paths.get(casesDir + slipPropertyFilename(slipOpinion.getFileName(), mainCaseScreen))) 
 			) {
-				slipOpinion.setSummary( parseMainCaseScreenDetail(inputStream) ); 
+				goodtogo = parseMainCaseScreenDetail(inputStream, slipOpinion); 
 			} catch (IOException e) {
-				slipOpinion.setSummary( new Summary() ); 
-				System.out.println("File error: " + e.getMessage());
+				logger.warning("File error: " + e.getMessage());
+				goodtogo = false;
 			}
+			if ( !goodtogo ) {
+				continue;
+			}
+			
 			try ( InputStream inputStream = Files.newInputStream( 
 					Paths.get(casesDir + slipPropertyFilename(slipOpinion.getFileName(), disposition))) 
 			) {
-				slipOpinion.setDisposition( parseDispositionDetail(inputStream) ); 
+				parseDispositionDetail(inputStream, slipOpinion); 
 			} catch (IOException e) {
-				slipOpinion.setDisposition( new Disposition() ); 
-				System.out.println("File error: " + e.getMessage());
+				logger.warning("File error: " + e.getMessage());
 			}
 			try ( InputStream inputStream = Files.newInputStream( 
 					Paths.get(casesDir + slipPropertyFilename(slipOpinion.getFileName(), partiesAndAttorneys))) 
 			) {
-				slipOpinion.setPartiesAndAttornies( parsePartiesAndAttorneysDetail(inputStream) ); 
+				parsePartiesAndAttorneysDetail(inputStream, slipOpinion); 
 			} catch (IOException e) {
-				slipOpinion.setPartiesAndAttornies( new PartiesAndAttornies() ); 
-				System.out.println("File error: " + e.getMessage());
+				logger.warning("File error: " + e.getMessage());
 			}
 			try ( InputStream inputStream = Files.newInputStream( 
 					Paths.get(casesDir + slipPropertyFilename(slipOpinion.getFileName(), trialCourt))) 
 			) {
-				slipOpinion.setTrialCourt( parseTrialCourtDetail(inputStream) ); 
+				parseTrialCourtDetail(inputStream, slipOpinion); 
 			} catch (IOException e) {
-				slipOpinion.setTrialCourt( new TrialCourt() ); 
-				System.out.println("File error: " + e.getMessage());
+				logger.warning("File error: " + e.getMessage());
 			}
-			
+/*			
+			String caption = slipOpinion.getSummary().getCaseCaption();
+			if ( caption != null )
+				slipOpinion.setTitle(caption);
+			String division = slipOpinion.getSummary().getDivision();
+			if ( division != null )
+				slipOpinion.setCourt(division );
+			Date dispositionDate = slipOpinion.getDisposition().getDate();
+			if ( dispositionDate != null )
+				slipOpinion.setOpinionDate(dispositionDate);
+*/					
 		}		
 		return documents;
