+++ OpinionsEjb/src/main/java/opca/service/CAOnlineUpdates.java
@@ -157,7 +157,8 @@ public class CAOnlineUpdates {
 
 		processOpinions(citationStore, mergeOpinions, persistOpinions);
 	  	processStatutes(citationStore, mergeStatutes, persistStatutes);
-		
+		int cSlip = 0;
+				
 		List<OpinionStatuteCitation> persistOpinionStatuteCitations = new ArrayList<>();
 		for( SlipOpinion slipOpinion: slipOpinions ) {
 			if ( slipOpinion.getStatuteCitations() != null ) {
@@ -166,8 +167,13 @@ public class CAOnlineUpdates {
 	    		}
 			}
 			em.persist(slipOpinion);
+			if ( slipOpinion.getSlipProperties() == null ) {
+				System.out.println("SlipProperties == null " );
+			}
 			em.persist(slipOpinion.getSlipProperties());
+			cSlip++;
 		}
+		System.out.println("cSlip = " + cSlip);
 		Date startTime = new Date();
     	for(OpinionBase opinion: persistOpinions ) {
 			em.persist(opinion);
