+++ OpinionsWeb/src/main/webapp/resources/opview/opview.xhtml
@@ -53,7 +53,7 @@ xmlns:composite="http://xmlns.jcp.org/jsf/composite">
         <ui:fragment rendered="#{cc.attrs.view.cases.size() > 10}"><div class="opencase casehead"><span class="opencase case">[<h:outputText value="#{cc.attrs.view.cases.size() - 10}"/> more cases cited.]</span></div></ui:fragment>
       </div>
      </ui:fragment>
-     <ui:fragment rendered="#{cc.attrs.view.disposition != null}" ><div class="summary wrap"><div class="summary outer"><div class="summary inner"><b><h:outputText value="#{cc.attrs.view.disposition}: "/></b><h:outputText value="#{cc.attrs.view.summary}" /></div></div></div></ui:fragment>
+     <ui:fragment rendered="#{cc.attrs.view.disposition != null}" ><div class="summary wrap"><div class="summary outer"><div class="summary inner"><b><h:outputText value="#{cc.attrs.view.disposition}: "/></b><h:outputText value="Summary Placeholder" /></div></div></div></ui:fragment>
     <div class="opinion tail" />
     </h:form>
   </div>
